+++ api/include/opentelemetry/trace/default_span.h
@@ -48,7 +48,10 @@ public:
   {}
 
 #if OPENTELEMETRY_ABI_VERSION_NO >= 2
-  void AddLink(const SpanContextKeyValueIterable * /* links */) noexcept override {}
+  void AddLink(const SpanContext & /* target */,
+               const common::KeyValueIterable & /* attrs */) noexcept override
+  {}
+  void AddLinks(const SpanContextKeyValueIterable & /* links */) noexcept override {}
 #endif
 
   void SetStatus(StatusCode /* status */, nostd::string_view /* description */) noexcept override {}