+++ sdk/src/trace/span.h
@@ -46,7 +46,10 @@ public:
                  nostd::string_view description) noexcept override;
 
 #if OPENTELEMETRY_ABI_VERSION_NO >= 2
-  void AddLink(const opentelemetry::trace::SpanContextKeyValueIterable *links) noexcept override;
+  void AddLink(const opentelemetry::trace::SpanContext &target,
+               const opentelemetry::common::KeyValueIterable &attrs) noexcept override;
+
+  void AddLinks(const opentelemetry::trace::SpanContextKeyValueIterable &links) noexcept override;
 #endif
 
   void UpdateName(nostd::string_view name) noexcept override;